feat: Complete M15-M41 implementation — full institutional trading system
All milestones from the lively-kindling-lovelace plan are now implemented:
- M15-M18: Features (fundamental, behavioral, buyback, supply chain, analyst, seasonal, short interest, index rebal, cross-asset, NLP, institutional, VPIN)
- M19: ML (temporal attention, GNN, symbolic regression, mutual info selection, online ensemble, AutoML, Bayesian NN, conformal prediction, causal inference, calibration, MAML, TDA, RL, Gaussian process, graph models)
- M20: Execution (adaptive algo, Almgren-Chriss, SOR, IBKR adapter, pre-open signals, intraday features)
- M21: Risk (intraday monitor, tail hedging, stressed VaR, regime costs, reverse stress)
- M23: Streaming (WebSocket client, minute bar aggregator)
- M25-M26: Signals (behavioral finance, stat arb, strategy discovery)
- M27: Portfolio (multi-period, risk budgeting, robust optimizer, strategy allocator, cost-aware enhancements)
- M28: Compliance (audit log, OTR monitor, regulatory reports)
- M29-M30: Ops (dashboard data, certification, self-healing, event bus)
- M31: Grafana dashboards
- M32: E2E integration testing
- M36: QA (A/B testing, adversarial testing, capacity, multiple testing)
- M37: Factor timing (value spread, macro conditioning)
- M38: Tail risk hedging
- M39: Regime-conditional transaction costs
- M40: A/B testing framework

1243 phase12 tests passing, 0 failures, 8 skipped.

+++ src/assembled_core/features/market_breadth.py
@@ -539,7 +539,7 @@ def compute_zweig_breadth_thrust(
     # Thrust signal: ema crosses from < 0.40 to > 0.615 within window days
     below_threshold = (ema < 0.40).astype(int)
     above_threshold = (ema > 0.615).astype(int)
-    had_low_recently = below_threshold.rolling(window, min_periods=1).max()
+    had_low_recently = below_threshold.rolling(window, min_periods=1).max().astype(int)
     thrust_signal = (above_threshold & had_low_recently).astype(float)
 
     result = pd.DataFrame({
